add allergen list in user profile, refactor to use fragments
