+++ demo/src/main/java/io/weet/demo/models/Allergen.java
@@ -1,15 +1,11 @@
-package io.weet.demo.controllers;
-
-
+package io.weet.demo.models;
 
 public class Allergen {
-   // private String user;
     private int id;
     private String allergen;
 
-    public Allergen(){
+    public Allergen(){}
 
-    }
     public Allergen(int id, String allergen){
         this.id=id;
         this.allergen=allergen;
@@ -26,13 +22,9 @@ public class Allergen {
     public void setAllergen(String allergen){
         this.allergen=allergen;
     }
-   /*public String getUser(){
-        return user;
-    }
-    public void setUser(String user){
-        this.user=user;
-    }*/
+
     @Override
     public String toString(){
-        return "id= "+ id + "allergen= "+allergen;   }
+        return String.format("ID = %d; Allergen = %s", id, allergen);
+    }
 }
