--- demo/src/main/resources/templates/Allergen.html
@@ -1,37 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="https://www.thymeleaf.org">
-    <head> 
-        <title>Home Page</title>
-        <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <h3> Welcome to Your Home Page!</h3>
-    <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    </head>
-    
-    <body>
-        <div class="container"></div>
-        
-            <table class="table table-bordered table-striped">
-                <thead class="thead-dark">
-                    <tr>
-                        <th>Allergens</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <tr th:each="tempAllergens:${allergens}">
-                        <td th:text="${tempAllergens.getId()}"></td>
-                        <td th:text="${tempAllergens.getAllergen()}" ></td>
-                       
-                    </tr>
-                </tbody>
-                
-         </table>
-         <div class="container text-center" style="margin-top: 2%;">
-            <button type="submit" class="btn btn-success">Add Allergen</button>
-            </div>
-        </div>
-    </body>
-
-</html>
