+++ demo/src/main/resources/templates/userProfile.html
@@ -1,0 +1,13 @@
+<!DOCTYPE HTML>
+<html xmlns:th="https://www.thymeleaf.org">
+    <head th:replace ="fragments/head::head"></head>
+    <body>
+        <div th:replace ="fragments/navigator::nav"></div>
+        <main style="margin-top: 100px;">
+            <div>
+                <h1 class="display-4 text-center"> User Profile</h1>
+                <div th:replace ="fragments/allergen::allergenList(allergens=${allergenList})"></div>
+            </div>
+        </main>
+    </body>
+</html>
